feat: create individual page for dataset in URL of /data/{id}, and added link in /data page table
