+++ sites/geohub/src/routes/(app)/data/[id]/+page.ts
@@ -1,0 +1,2 @@
+export const csr = false;
+export const ssr = true;
